sway-window-move: allow creation of workspaces on window move
this commit now allows a window to be moved to a non-existent workspace
by specifying a non-matching string in the fzf window.

if the workspace name fuzzy matches an existing workspace you may suffix
the workspace name with ":new" to break the match; the script will rip
off the ":new" portion and create the workspace.

this commit also adds prompts to each fzf window created by each script.

Signed-off-by: ldelossa <louis.delos@gmail.com>